divide now returns double

+++ src/main/java/com/example/springdocker/MyMathCalc/MyMathCalc.java
@@ -10,7 +10,7 @@ public class MyMathCalc {
         return a * b;
     }
 
-    public float divide(float a, float b) {
+    public double divide(double a, double b) {
         return a / b;
     }
 }
